Refactor pow_syndacoin.py
