+++ powcoin/pow_syndacoin.py
@@ -19,13 +19,13 @@ from copy import deepcopy
 from ecdsa import SigningKey, SECP256k1
 from utils import serialize, deserialize
 
-from identities import user_private_key, user_public_key, node_public_key
+from identities import user_private_key, user_public_key
 
 
 PORT = 10000
 node = None
 
-logging.basicConfig(level="INFO", format="%(threadName)-6s | %(message)s")
+logging.basicConfig(level="INFO", format='%(threadName)-6s | %(message)s')
 logger = logging.getLogger(__name__)
 
 
@@ -75,25 +75,25 @@ class TxOut:
 
 class Block:
 
-    def __init__(self, txns, prev_id, nonce=0):
+    def __init__(self, txns, prev_id, nonce):
         self.txns = txns
         self.prev_id = prev_id
         self.nonce = nonce
+
+    @property
+    def header(self):
+        return serialize(self)
 
     @property
     def id(self):
         return hashlib.sha256(self.header).hexdigest()
 
     @property
-    def header(self):
-        return serialize([self.txns, self.prev_id, self.nonce])
-
-    @property
     def proof(self):
         return int(self.id, 16)
 
     def __repr__(self):
-        return f"Block(prev_id={self.prev_id}, id={self.id} nonce={self.nonce})"
+        return f"Block(prev_id={self.prev_id[:10]}... id={self.id[:10]}...)"
 
 class Node:
 
@@ -153,35 +153,34 @@ class Node:
         # Check no value created or destroyed
         assert in_sum == out_sum
 
-    def validate_block(self, block):
-        assert int(block.id, 16) < POW_TARGET, "Insufficient Proof-of-Work"
-        assert block.prev_id == self.blocks[-1].id
-
     def handle_tx(self, tx):
         self.validate_tx(tx)
         self.mempool.append(tx)
 
+    def validate_block(self, block):
+        assert block.proof < POW_TARGET, "Insufficient Proof-of-Work"
+        assert block.prev_id == self.blocks[-1].id
+
     def handle_block(self, block):
-        with chain_lock:
-            # Check the proof-of-work
-            self.validate_block(block)
+        # Check work, chain ordering
+        self.validate_block(block)
 
-            # Check the transactions are valid
-            for tx in block.txns:
-                self.validate_tx(tx)
+        # Check the transactions are valid
+        for tx in block.txns:
+            self.validate_tx(tx)
 
-            # If they're all good, update self.blocks and self.utxo_set
-            for tx in block.txns:
-                self.update_utxo_set(tx)
-            
-            # Add the block and increment the id of bank who will report next block
-            self.blocks.append(block)
+        # If they're all good, update self.blocks and self.utxo_set
+        for tx in block.txns:
+            self.update_utxo_set(tx)
+        
+        # Add the block to our chain
+        self.blocks.append(block)
 
-            logging.info(f"Block accepted: height={len(self.blocks) - 1}")
+        logger.info(f"Block accepted: height={len(self.blocks) - 1}")
 
-            # Tell peers
-            for peer in self.peer_addresses:
-                send_message(peer, "block", block)
+        # Block propogation
+        for peer_address in self.peer_addresses:
+            send_message(peer_address, "block", block)
 
 def prepare_simple_tx(utxos, sender_private_key, recipient_public_key, amount):
     sender_public_key = sender_private_key.get_verifying_key()
@@ -211,46 +210,48 @@ def prepare_simple_tx(utxos, sender_private_key, recipient_public_key, amount):
     for i in range(len(tx.tx_ins)):
         tx.sign_input(i, sender_private_key)
 
-    return tx
-
 ##########
 # Mining #
 ##########
 
-BITS = 20
-POW_TARGET = 2 ** (256 - BITS)
-
+DIFFICULTY_BITS = 20
+POW_TARGET = 2 ** (256 - DIFFICULTY_BITS)
 mining_interrupt = threading.Event()
-chain_lock = threading.Lock()
 
 
-def mine_block(block, rand_step=True):
-    step = random.randint(0, 1000) if rand_step else 1
+def mine_block(block):
     while block.proof >= POW_TARGET:
-        block.nonce += step
+        # TODO: accept interrupts here if tip changes
         if mining_interrupt.is_set():
             logger.info("Mining interrupted")
             mining_interrupt.clear()
             return
+        block.nonce += 1
     return block
 
 
-def mine_forever(public_key):
+def mine_forever():
     logging.info("Starting miner")
     while True:
-        with chain_lock:
-            unmined_block = Block(
-                txns=node.mempool,
-                prev_id=node.blocks[-1].id,
-                nonce=random.randint(0, 1000),
-            )
+        unmined_block = Block(
+            txns=node.mempool,
+            prev_id=node.blocks[-1].id,
+            nonce=random.randint(0, 1000000000),
+        )
         mined_block = mine_block(unmined_block)
-        
-        # This is False if mining was interrupted
+
         if mined_block:
             logger.info("")
-            logger.info(f"Mined block")
+            logger.info("Mined a block")
             node.handle_block(mined_block)
+
+def mine_genesis_block():
+    global node
+    unmined_block = Block(txns=[], prev_id=None, nonce=0)
+    mined_block = mine_block(unmined_block)
+    node.blocks.append(mined_block)
+    # TODO: update utxo set, award coinbase, etc
+
 
 ##############
 # Networking #
@@ -280,9 +281,9 @@ class TCPHandler(socketserver.BaseRequestHandler):
             self.respond(command="pong", data="")
 
         if command == "block":
-            # If it extends the tip
             if data.prev_id == node.blocks[-1].id:
                 node.handle_block(data)
+                # Interrupt mining thread
                 mining_interrupt.set()
 
         if command == "tx":
@@ -302,6 +303,7 @@ def external_address(node):
     return ('localhost', port)
 
 def serve():
+    logger.info("Starting server")
     server = socketserver.TCPServer(("0.0.0.0", PORT), TCPHandler)
     server.serve_forever()
 
@@ -313,35 +315,25 @@ def send_message(address, command, data, response=False):
         if response:
             return deserialize(s.recv(5000))
 
-def mine_genesis_block():
-    unmined_genesis_block = Block(txns=[], prev_id=None)
 
-    # We hard-code nonce and step so everyone mines the same coinbase
-    mined_genesis_block = mine_block(unmined_genesis_block, rand_step=False)
-
-    node.blocks.append(mined_genesis_block)
-    return node
+#######
+# CLI #
+#######
 
 def main(args):
     if args["serve"]:
-        # For logging purposes, rename main thread
-        threading.current_thread().name = "main"
-
         global node
         node = Node()
-
-        # Mine the genesis block
+        
+        # TODO: mine genesis block
         mine_genesis_block()
 
-        # First thing, start server in one thread thread
+        # Start server thread
         server_thread = threading.Thread(target=serve, name="server")
         server_thread.start()
 
-        # First thing, start miner in one thread thread
-        logger.info("starting miner")
-        node_id = int(os.environ["ID"])
-        mining_public_key = node_public_key(node_id)
-        miner_thread = threading.Thread(target=mine_forever, args=(mining_public_key,), name="miner")
+        # Start miner thread
+        miner_thread = threading.Thread(target=mine_forever, name="miner")
         miner_thread.start()
 
     elif args["ping"]:
